# CC2538: Faster crypto