This PR brings faster re-implementations of AES-128 and CCM*. Executing Test vector C.1 from FIPS Pub 197 one thousand times took 49.3ms with the old and 25.75ms with the new AES-128 driver. Executing Test vector C.2.1.2 from IEEE 802.15.4-2006 one thousand times took 94.60ms with the old and 59.75ms with the new CCM* driver. My CSL implementation depends on these faster timings.

A known issue with both the old and new CCM* driver for the CC2538 is that it freezes on certain inputs (e.g., a_len = 0 and a_len = 0 & m_len = 0). I fix for that will come later.